Extract process snapshot collector
